+++ spec/controllers/application_controller_spec.rb
@@ -1,0 +1,5 @@
+require 'rails_helper'
+
+RSpec.describe 'test example' do
+  specify { expect(true).to eq(true) }
+end