+++ spec/rails_helper.rb
@@ -1,0 +1,47 @@
+require 'spec_helper'
+
+ENV['RAILS_ENV'] ||= 'test'
+
+require File.expand_path('../config/environment', __dir__)
+abort('The Rails environment is running in production mode!') if Rails.env.production?
+
+require 'rspec/rails'
+
+FactoryBot.definition_file_paths = [File.expand_path('factories', __dir__)]
+FactoryBot.find_definitions
+
+RSpec.configure do |config|
+  config.include FactoryBot::Syntax::Methods
+  config.infer_spec_type_from_file_location!
+  config.filter_rails_from_backtrace!
+  config.use_transactional_fixtures = false
+
+  config.around(:each) do |example|
+    DatabaseCleaner.cleaning do
+      example.run
+    end
+  end
+
+  config.before { allow($stdout).to receive(:puts) }
+  config.before(:suite) do
+    Rails.application.load_tasks
+    DatabaseCleaner.strategy = :transaction
+    DatabaseCleaner.clean_with(:truncation)
+  end
+
+  # Tag Rake specs with `:task` metadata or put them in the spec/tasks dir
+  config.shared_context_metadata_behavior = :apply_to_host_groups
+  config.define_derived_metadata(:file_path => %r{/spec/tasks/}) do |metadata|
+    metadata[:type] = :task
+  end
+end
+
+require 'shoulda/matchers'
+Shoulda::Matchers.configure do |config|
+  config.integrate do |with|
+    with.test_framework :rspec
+    with.library :rails
+  end
+end
+
+ActiveRecord::Migration.maintain_test_schema!