+++ CHANGELOG.md
@@ -3,6 +3,10 @@
 ## HEAD Unreleased
 ### Latest update: 2018-11-22
 
+### Feature Request
+
+- [PR #2](https://github.com/mikeyuchima/PitCrew-Rails-API/pull/2) Add Rails API structure and test suite [@vbrazo](https://github.com/vbrazo)
+
 ### Documentation
 
 - [PR #1](https://github.com/mikeyuchima/PitCrew-Rails-API/pull/1) Set up project [@vbrazo](https://github.com/vbrazo)